+++ dashboard/dashboard-front/src/components/jobs/JobsListView.vue
@@ -1,11 +1,15 @@
 <script setup lang="ts">
-import { computed, nextTick, onMounted, ref, watch, type Ref } from 'vue'
+import { computed, nextTick, onMounted, onUnmounted, ref, watch, type Ref } from 'vue'
 import { QTree } from 'quasar'
+import { io, Socket } from "socket.io-client"
+import { mdiDotsVertical } from '@quasar/extras/mdi-v7'
 import { apiClient } from '@/services/ApiClient'
 import { toastService } from '@/services/ToastService'
+import { envInfo } from '@/services/EnvironmentInfo'
 import { type JobData } from '@/utils/api-schema'
 import JobDetails from '@/components/jobs/JobDetails.vue'
 import JobStatus from '@/components/jobs/JobStatus.vue'
+import TimeAgoLabel from '@/components/jobs/TimeAgoLabel.vue'
 import { JobOrder, filterJobByKeyword, sortedJobs } from '@/utils/jobs'
 
 const jobsData: Ref<JobData[]> = ref([])
@@ -19,10 +23,13 @@ const jobsCount: Ref<number> = computed(() => jobsData.value?.length || 0)
 const selectedNodeKey: Ref<string | null> = ref(null)
 const jobOrder: Ref<JobOrder> = ref(JobOrder.ByName)
 const loadingTree = ref(true)
+const autoUpdateEnabled = ref(true)
+const lastReloadTimestamp: Ref<number> = ref(0)
 
 function fetchJobs() {
     loadingTree.value = true
     apiClient.get(`/api/v1/job`).then(response => {
+        lastReloadTimestamp.value = new Date().getTime() / 1000
         jobsData.value = response.data
     }).catch(err => {
         toastService.showErrorDetails(`Failed to fetch the jobs`, err)
@@ -131,6 +138,47 @@ watch(jobsTree, () => {
     })
 })
 
+watch(envInfo, () => {
+    setupEventStreamClient()
+})
+watch(autoUpdateEnabled, () => {
+    localStorage.setItem('jobs.autoUpdateEnabled', JSON.stringify(autoUpdateEnabled.value))
+    setupEventStreamClient()
+})
+
+var autoReloadSocket: Socket | null = null
+
+function setupEventStreamClient() {
+    autoReloadSocket?.disconnect()
+    autoReloadSocket = null
+
+    if (!envInfo.lifecycle_url)
+        return
+    if (!autoUpdateEnabled.value)
+        return
+
+    // Socket.io needs URL without the path
+    const url = new URL(envInfo.lifecycle_url)
+    url.pathname = ''
+    const trimmedLifecycleUrl = url.toString()
+    const socket = io(trimmedLifecycleUrl, {
+        path: '/lifecycle/socketio/events',
+    })
+    autoReloadSocket = socket
+
+    socket.on("connect", () => {
+        console.log(`connected to live events stream: ${socket.id}`)
+        socket.on("broadcast_event", (data) => {
+            console.log('Change detected, reloading jobs')
+            fetchJobs()
+        })
+    })
+
+    socket.on("disconnect", () => {
+        console.log('disconnected from live events stream')
+    })
+}
+
 onMounted(() => {
     const storedOrder = localStorage.getItem('jobs.order')
     if (storedOrder) {
@@ -141,7 +189,21 @@ onMounted(() => {
         }
     }
 
+    const storedAutoUpdateEnabled = localStorage.getItem('jobs.autoUpdateEnabled')
+    if (storedAutoUpdateEnabled) {
+        try {
+            autoUpdateEnabled.value = JSON.parse(storedAutoUpdateEnabled) as boolean
+        } catch(e) {
+            console.error(e)
+        }
+    }
+
     fetchJobs()
+    setupEventStreamClient()
+})
+
+onUnmounted(() => {
+    autoReloadSocket?.disconnect()
 })
 </script>
 <template>
@@ -193,6 +255,24 @@ onMounted(() => {
                             </q-list>
                         </q-btn-dropdown>
                         </span>
+
+                        <q-btn-dropdown rounded flat color="grey-7" :dropdown-icon="mdiDotsVertical" no-icon-animation>
+                            <q-list>
+                                <q-item>
+                                    <span class="q-mt-sm">
+                                        Last update:
+                                        <TimeAgoLabel :timestamp="lastReloadTimestamp" />
+                                    </span>
+                                </q-item>
+                                <q-item>
+                                    <q-toggle v-model="autoUpdateEnabled" label="Auto-update">
+                                        <q-tooltip anchor="center right" self="center left">
+                                            Refresh jobs in real time as soon as the change is detected.
+                                        </q-tooltip>
+                                    </q-toggle>
+                                </q-item>
+                            </q-list>
+                        </q-btn-dropdown>
 
                         <q-scroll-area style="height: 64vh;" visible>
                         <q-tree