+++ lifecycle/lifecycle/server/metrics.py
@@ -23,6 +23,16 @@ def setup_lifecycle_metrics():
     REGISTRY.register(DatabaseConnectionCollector())
 
 
+def unregister_metrics():
+    """
+    Clean up Prometheus metrics registered by this app.
+    It's important if you want to run the server more than once, eg. in tests.
+    """
+    collectors = list(REGISTRY._collector_to_names.keys())
+    for collector in collectors:
+        REGISTRY.unregister(collector)
+
+
 class DatabaseConnectionCollector(Collector):
     def collect(self):
         metric_name = 'lifecycle_database_connected'