+++ lifecycle/tests/server/test_version.py
@@ -1,15 +1,19 @@
 import os
 
+import pytest
 from fastapi.testclient import TestClient
 
 from lifecycle.config import Config
+from lifecycle.server.metrics import unregister_metrics
 from racetrack_commons.plugin.engine import PluginEngine
 from lifecycle.server.api import create_fastapi_app
 
 
+@pytest.mark.django_db(transaction=True)
 def test_health_version_endpoint():
     os.environ['GIT_VERSION'] = '0.0.1-g32c4b29-dirty'
     os.environ['DJANGO_DB_TYPE'] = 'sqlite'
+    unregister_metrics()
     fastapi_app = create_fastapi_app(Config(), PluginEngine(), 'lifecycle')
 
     client = TestClient(fastapi_app)