+++ racetrack_commons/racetrack_commons/api/asgi/asgi_server.py
@@ -72,6 +72,8 @@ def _setup_uvicorn_logs(access_log: bool):
     }
 
     LOGGING_CONFIG["loggers"]["uvicorn"]["propagate"] = False
+    LOGGING_CONFIG["loggers"]["uvicorn.error"]["level"] = 'INFO'
+    LOGGING_CONFIG["loggers"]["uvicorn.error"]["propagate"] = False
 
     if not access_log:
         LOGGING_CONFIG["loggers"]["uvicorn.access"]["level"] = 'CRITICAL'