+++ docs/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Jobs tree on the Dashboard is refreshed in real time as soon as someone else's change is detected.
+  You can turn it off on the "Jobs" page by clicking "3 dots menu", "Auto-update" toggle.
+  [issue #239](https://github.com/TheRacetrack/racetrack/issues/239)
 
 ## [2.15.0] - 2023-05-26
 ### Changed